backend/requirements.txt
# Core Flask dependencies
Flask==3.0.0
Werkzeug==3.0.1

# Data processing and analysis
pandas==2.1.4
numpy==1.26.4

# Machine Learning
scikit-learn==1.6.1
tensorflow==2.19.0
joblib==1.3.2

# Visualization
plotly==5.17.0

# File handling and utilities
h5py==3.12.1
openpyxl==3.1.2
xlrd==2.0.1

# Environment and configuration
python-dotenv==1.0.0

# Production WSGI server
gunicorn==21.2.0

# System utilities
psutil==6.1.0
python-dateutil==2.9.0.post0

# HTTP requests
requests==2.31.0
urllib3==2.2.3

# Additional utilities
colorama==0.4.6
click==8.1.8


supabase==2.18.1
python-dotenv==1.0.0
flask-login==0.6.3
bcrypt==4.1.2
flask_cors
